workers/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.1
minio==7.2.0
qdrant-client==1.7.0
sentence-transformers==2.3.1
torch==2.0.1
pypdf==3.17.0
pandas==2.1.3
pyarrow==14.0.1
python-dotenv==1.0.0
